refactor edgetpu to accept more detector types
