+++ frigate/detection/__init__.py
@@ -1,0 +1,114 @@
+import logging
+import numpy as np
+import multiprocessing as mp
+from frigate.util import EventsPerSecond
+from frigate.config import DetectorConfig, DetectorTypeEnum
+
+logger = logging.getLogger(__name__)
+
+
+class DetectionProcess:
+    def __init__(
+        self,
+        name,
+        detection_queue,
+        out_events,
+        model_path,
+        model_shape,
+        detector_config: DetectorConfig,
+    ):
+        self.name = name
+        self.out_events = out_events
+        self.detection_queue = detection_queue
+        self.avg_inference_speed = mp.Value("d", 0.01)
+        self.detection_start = mp.Value("d", 0.0)
+        self.detect_process = None
+        self.model_path = model_path
+        self.model_shape = model_shape
+        self.detector_config = detector_config
+
+        self.detector_target = None
+        if (
+            detector_config.type == DetectorTypeEnum.cpu
+            or detector_config.type == DetectorTypeEnum.edgetpu
+        ):
+            from .edgetpu import run_detector as edgetpu_detector
+
+            self.detector_target = edgetpu_detector
+
+        assert self.detector_target, "Invalid detector configuration"
+
+        self.start_or_restart()
+
+    def stop(self):
+        self.detect_process.terminate()
+        logging.info("Waiting for detection process to exit gracefully...")
+        self.detect_process.join(timeout=30)
+        if self.detect_process.exitcode is None:
+            logging.info("Detection process didnt exit. Force killing...")
+            self.detect_process.kill()
+            self.detect_process.join()
+
+    def start_or_restart(self):
+        self.detection_start.value = 0.0
+        if (not self.detect_process is None) and self.detect_process.is_alive():
+            self.stop()
+        self.detect_process = mp.Process(
+            target=self.detector_target,
+            name=f"detector:{self.name}",
+            args=(
+                self.name,
+                self.detection_queue,
+                self.out_events,
+                self.avg_inference_speed,
+                self.detection_start,
+                self.model_path,
+                self.model_shape,
+                self.detector_config,
+            ),
+        )
+        self.detect_process.daemon = True
+        self.detect_process.start()
+
+
+class RemoteObjectDetector:
+    def __init__(self, name, labels, detection_queue, event, model_shape):
+        self.labels = labels
+        self.name = name
+        self.fps = EventsPerSecond()
+        self.detection_queue = detection_queue
+        self.event = event
+        self.shm = mp.shared_memory.SharedMemory(name=self.name, create=False)
+        self.np_shm = np.ndarray(
+            (1, model_shape[0], model_shape[1], 3), dtype=np.uint8, buffer=self.shm.buf
+        )
+        self.out_shm = mp.shared_memory.SharedMemory(
+            name=f"out-{self.name}", create=False
+        )
+        self.out_np_shm = np.ndarray((20, 6), dtype=np.float32, buffer=self.out_shm.buf)
+
+    def detect(self, tensor_input, threshold=0.4):
+        detections = []
+
+        # copy input to shared memory
+        self.np_shm[:] = tensor_input[:]
+        self.event.clear()
+        self.detection_queue.put(self.name)
+        result = self.event.wait(timeout=10.0)
+
+        # if it timed out
+        if result is None:
+            return detections
+
+        for d in self.out_np_shm:
+            if d[1] < threshold:
+                break
+            detections.append(
+                (self.labels[int(d[0])], float(d[1]), (d[2], d[3], d[4], d[5]))
+            )
+        self.fps.update()
+        return detections
+
+    def cleanup(self):
+        self.shm.unlink()
+        self.out_shm.unlink()
