+++ frigate/detection/object_detector.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class ObjectDetector(ABC):
+    @abstractmethod
+    def detect(self, tensor_input, threshold=0.4):
+        pass
