+++ frigate/util.py
@@ -602,6 +602,25 @@ def add_mask(mask, mask_img):
     cv2.fillPoly(mask_img, pts=[contour], color=(0))
 
 
+def load_labels(path, encoding="utf-8"):
+    """Loads labels from file (with or without index numbers).
+    Args:
+      path: path to label file.
+      encoding: label file encoding.
+    Returns:
+      Dictionary mapping indices to labels.
+    """
+    with open(path, "r", encoding=encoding) as f:
+        lines = f.readlines()
+        if not lines:
+            return {}
+
+        if lines[0].split(" ", maxsplit=1)[0].isdigit():
+            pairs = [line.split(" ", maxsplit=1) for line in lines]
+            return {int(index): label.strip() for index, label in pairs}
+        else:
+            return {index: line.strip() for index, line in enumerate(lines)}
+
 class FrameManager(ABC):
     @abstractmethod
     def create(self, name, size) -> AnyStr:
